♻️ REFACTOR: verdi export/import -> verdi archive
Deprecated groups/commands are hidden (so they don't show in `verdi --help`) and decorated with `deprecated_command`.
